requirements.txt
azure-functions
pymongo
azure-ai-formrecognizer         
python-docx
requests
openai
dnspython
azure-storage-blob